refactor(cache): split connection from schema
