+++ packages/core/src/discovery/cache/model-cost.ts
@@ -6,7 +6,7 @@ import type { ModelCostEntry } from "../../contract/index.js";
 import type { ProjectIdentityKind } from "../../types/index.js";
 import type { DatabaseRow, SQLiteDatabase } from "../../utils/sqlite.js";
 import { hasCacheStorage } from "./db.js";
-import { withCacheDbReadOnly } from "./schema.js";
+import { withCacheDbReadOnly } from "./connection.js";
 
 export type { ModelCostEntry };
