+++ packages/core/src/discovery/index.ts
@@ -43,7 +43,7 @@ export type {
   SaveCachedSessionsOptions,
   SessionSnapshotCompleteness,
 } from "./cache/sessions.js";
-export type { CacheReadOutcome } from "./cache/schema.js";
+export type { CacheReadOutcome } from "./cache/connection.js";
 export { closeCacheStorage, getCachePath } from "./cache/db.js";
 export type { PersistedSessionHeadChange } from "./cache/db.js";
 export { getAnalyticsRevision } from "./cache/analytics-revision.js";
